refactor(workflow): extract config validation into validate.go, decomposed (#410)
validateConfig was a 190-line, gocognit-~50 monolith and the #410 burn-down's
named hotspot #2. This change:
- splits validateConfig into a fixed-order orchestrator plus per-section
  validators (tracker/repo, Linear routes, services, supported values,
  sandbox, server port, codex/claude, agent limits, timeouts), each with a
  single nameable responsibility and well under the 80-line / gocognit-10
  budgets; and
- moves the whole validation family out of loader.go into a new validate.go,
  bringing loader.go back from 1071 to 744 lines (under the 800-line file
  budget per AGENTS.md clean-code rule 7) and folding a stale doc comment that
  described validateConfig but sat above hasExplicitServiceRoute into the
  orchestrator doc where it belongs.

The validator slice order preserves the original first-error precedence
exactly (tracker.kind first per SPEC §6.4), so every config produces a
byte-identical error — pure structural refactor, no behavior change.

Before splitting, add characterization tests for the two validateConfig
branches that had no coverage and are easiest to mis-extract:
- Linear service custom_fields rejection (#326), reachable via Load.
- the defensive codex.turn_sandbox_policy.Validate guard, reachable only via a
  programmatically constructed Config (UnmarshalYAML rejects bad types first),
  so it is pinned with a direct validateConfig call.

Both are mutation-verified (each fails when its production branch is neutered).
workflow-package funlen/gocognit findings drop 19 -> 15 with no validate*
function remaining a finding; statement coverage 78.3% -> 79.2%.
